Downgrade Golang to 1.12
First of all we mention that this project can be built with Go 1.12, so
we should make sure this is really the case. On the other side our
projects provide binaries which are also cross-compiled for arm64,
currently 1.13 does not build the arm64 binaries, so let's use a Go
version that is known to do that correctly.